feat: relocate body style into web
