+++ apps/storybook/.storybook/preview-head.html
@@ -1,0 +1,16 @@
+<style>
+  @font-face {
+    font-family: "MaruBuri";
+    src: url("https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-10-21@1.0/MaruBuri-Regular.woff")
+      format("woff2");
+  }
+  html {
+    font-family: "MaruBuri";
+  }
+  .docs-story {
+    height: fit-content;
+  }
+  .innerZoomElementWrapper {
+    height: 100vh;
+  }
+</style>
