--- apps/storybook/.storybook/preview.ts
@@ -1,40 +1,0 @@
-import type { Preview } from "@storybook/react";
-import { Body } from "@tripie/design-system";
-import maruBuri from "@tripie/design-system/components/typography/font";
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-  decorators: [
-    (Story, context) => {
-      return Body({
-        children: Story(),
-        ...context.args.style,
-        className: maruBuri.className,
-      });
-    },
-  ],
-  globalTypes: {
-    theme: {
-      description: "Global theme for components",
-      defaultValue: "light",
-      toolbar: {
-        // The label to show for this toolbar item
-        title: "Theme",
-        icon: "circlehollow",
-        // Array of plain string values or MenuItem shape (see below)
-        items: ["light", "dark"],
-        // Change title based on selected value
-        dynamicTitle: true,
-      },
-    },
-  },
-};
-
-export default preview;
