+++ apps/storybook/.storybook/preview.tsx
@@ -1,0 +1,42 @@
+import type { Preview } from "@storybook/react";
+import React from "react";
+import ThemeProvider from "../../web/provider/ThemeProvider";
+import "../../web/styles/page.scss";
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+
+  globalTypes: {
+    theme: {
+      description: "Global theme for components",
+      defaultValue: "light",
+      toolbar: {
+        // The label to show for this toolbar item
+        title: "Theme",
+        icon: "circlehollow",
+        // Array of plain string values or MenuItem shape (see below)
+        items: ["light", "dark"],
+        // Change title based on selected value
+        dynamicTitle: true,
+      },
+    },
+  },
+  decorators: [
+    (Story) => {
+      return (
+        <ThemeProvider>
+          <Story />
+        </ThemeProvider>
+      );
+    },
+  ],
+};
+
+export default preview;
