+++ apps/web/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 
 import Provider from "../provider/layout";
-
 export const metadata: Metadata = {
   title: "Tripie ✈️",
   description: "AI enhanced trip planner",
