+++ apps/web/app/page.tsx
@@ -10,6 +10,7 @@ import Container from "@tripie/design-system/components/container/_container";
 import UnstyledLink from "@tripie/design-system/components/typography/link/_link";
 import Paragraph from "@tripie/design-system/components/typography/paragraph/_paragraph";
 import ThemeButton from "../components/ThemeButton";
+import "../styles/page.scss";
 
 export default function Home() {
   const text =
