+++ apps/web/provider/layout.tsx
@@ -1,5 +1,11 @@
 import { ReactNode } from "react";
-import ThemeProvider from "./ThemeProvider";
+
+import dynamic from "next/dynamic";
+
+/**
+ * https://nextjs.org/docs/pages/building-your-application/optimizing/lazy-loading#with-no-ssr
+ */
+const ThemeProvider = dynamic(() => import("./ThemeProvider"), { ssr: false });
 
 export default function Provider({
   children,
