+++ packages/design-system/package.json
@@ -3,7 +3,6 @@
   "version": "0.0.0",
   "private": true,
   "exports": {
-    "./my-button": "./src/my-button.tsx",
     ".": [
       "./src/index.ts",
       "./src/index.tsx"
@@ -39,7 +38,6 @@
   },
   "dependencies": {
     "classnames": "^2.5.1",
-    "@next/font": "^13.5.6",
     "next": "15.0.0-rc.0",
     "postcss": "^8",
     "react": "^18.2.0",
