+++ packages/design-system/src/base/_font-stack.scss
@@ -4,7 +4,7 @@ $config-font-stack: () !default;
 $config-font-stack: map.deep-merge(
   (
     "default": (
-      "maruBuri",
+      "MaruBuri",
       "system-ui",
       "-apple-system",
       "Segoe UI",
