--- packages/design-system/src/components/body/_body.scss
@@ -1,13 +1,0 @@
-@use "../../base/" as *;
-@use "../../functions/" as *;
-@use "../../generator/" as *;
-@use "../../mixins/" as *;
-@use "./night-sky" as *;
-@use "sass:map";
-
-@include generate-styles;
-
-html[data-theme="dark"]:root {
-  color: color(default, 800);
-  @include night-sky;
-}
