--- packages/design-system/src/components/body/_body.stories.tsx
@@ -1,37 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import { useEffect } from "react";
-import { useAppTheme } from "../../hooks";
-import Body from "./_body";
-
-const meta: Meta<typeof Body> = {
-  title: "tripie-ui/Body",
-  tags: ["autodocs"],
-  component: Body,
-
-  decorators: [
-    (story, context) => {
-      const { mode, setMode } = useAppTheme();
-      const selectedTheme = context.globals.theme || mode;
-
-      useEffect(() => {
-        setMode(selectedTheme);
-      }, [selectedTheme]);
-
-      return <div className={`${context.globals.theme}`}>{story()}</div>;
-    },
-  ],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  name: "Default",
-  args: {
-    children: "default theme",
-    style: {
-      width: "100%",
-    },
-  },
-};
