--- packages/design-system/src/components/body/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./_body";
