+++ packages/design-system/src/components/index.ts
@@ -1,4 +1,3 @@
-export { default as Body } from "./body";
 export { default as Drawer } from "./drawer";
 export { default as MyButton } from "./myButton";
 export { default as Headers } from "./typography";
