--- packages/design-system/src/components/typography/font/index.ts
@@ -1,2 +1,0 @@
-export * from "./_font";
-export { default } from "./_font";
