+++ packages/design-system/src/custom/_night-sky.scss
@@ -1,0 +1,42 @@
+@use "../mixins/" as *;
+@use "../base/" as *;
+@use "sass:map";
+
+@mixin night-sky {
+  .background-container {
+    overflow-x: hidden;
+    // min-height: 100vh;
+    @include position(fixed);
+  }
+
+  .stars {
+    background: rgb(0, 0, 0) url("../static/images/stars.png") repeat;
+    @include position(absolute);
+    @include z-index(base);
+  }
+
+  @keyframes move-background {
+    from {
+      -webkit-transform: translate3d(0px, 0px, 0px);
+    }
+    to {
+      -webkit-transform: translate3d(1000px, 0px, 0px);
+    }
+  }
+
+  .twinkling {
+    background: transparent url("../static/images/twinkling.png") repeat;
+    @include position(absolute);
+    @include z-index(mask);
+    animation: move-background 70s linear infinite;
+    img {
+      @include position(absolute);
+      @include z-index(fixed);
+    }
+  }
+
+  .layout-wrap {
+    @include position(absolute);
+    @include z-index(masked);
+  }
+}
