+++ packages/design-system/src/custom/index.scss
@@ -1,0 +1,1 @@
+@forward "night-sky";
